# backend/src/accounts/adapters/inbound/http/views.py
"""Accounts HTTP views (inbound adapter) — auth + saved searches."""

from __future__ import annotations

from collections.abc import Mapping

from django.contrib.auth import get_user_model
from drf_spectacular.utils import extend_schema
from rest_framework import status
from rest_framework.exceptions import ValidationError
from rest_framework.permissions import AllowAny, IsAuthenticated
from rest_framework.request import Request
from rest_framework.response import Response
from rest_framework.throttling import ScopedRateThrottle
from rest_framework.views import APIView
from rest_framework_simplejwt.exceptions import TokenError
from rest_framework_simplejwt.tokens import RefreshToken

from accounts.adapters.inbound.http.serializers import (
    AccountSerializer,
    GoogleAuthRequestSerializer,
    LogoutRequestSerializer,
    RegisteredSerializer,
    RegisterSerializer,
    SavedSearchSerializer,
    TokenPairSerializer,
)
from accounts.adapters.outbound.persistence.models import ExternalIdentityModel
from accounts.application.errors import InvalidCredential
from accounts.composition import container
from catalog.adapters.inbound.http.serializers import VALIDATION_ERROR_SCHEMA, ErrorSerializer

User = get_user_model()


def _providers_for(user) -> list[str]:
    providers: list[str] = ["password"] if user.has_usable_password() else []
    for provider in (
        ExternalIdentityModel.objects.filter(user=user)
        .values_list("provider", flat=True)
        .distinct()
    ):
        if provider not in providers:
            providers.append(provider)
    return providers


class RegisterView(APIView):
    permission_classes = [AllowAny]
    throttle_classes = [ScopedRateThrottle]
    throttle_scope = "auth"

    @extend_schema(
        operation_id="auth_register",
        summary="Create an account",
        request=RegisterSerializer,
        responses={
            201: RegisteredSerializer,
            400: VALIDATION_ERROR_SCHEMA,
            429: ErrorSerializer,
        },
    )
    def post(self, request: Request) -> Response:
        serializer = RegisterSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        username = serializer.validated_data["username"]
        if User.objects.filter(username=username).exists():
            raise ValidationError({"username": "already taken"})
        user = User.objects.create_user(
            username=username, password=serializer.validated_data["password"]
        )
        return Response({"id": user.id, "username": user.username}, status=201)


class GoogleAuthView(APIView):
    """Exchange a Google ID token for our JWT, provisioning the account if new."""

    permission_classes = [AllowAny]
    throttle_classes = [ScopedRateThrottle]
    throttle_scope = "auth"

    @extend_schema(
        operation_id="auth_google",
        summary="Exchange a Google ID token for a session",
        request=GoogleAuthRequestSerializer,
        responses={
            200: TokenPairSerializer,
            400: VALIDATION_ERROR_SCHEMA,
            401: ErrorSerializer,
        },
    )
    def post(self, request: Request) -> Response:
        # `request.data` is whatever the client sent: a JSON array or scalar
        # gives `.get` an AttributeError, and this endpoint is unauthenticated,
        # so the resulting 500 is reachable without credentials.
        data = request.data if isinstance(request.data, Mapping) else {}
        credential = data.get("id_token")
        if not credential:
            raise ValidationError({"id_token": "required"})
        try:
            account = container.build_authenticate_with_google().execute(credential)
        except InvalidCredential:
            return Response({"detail": "Не удалось подтвердить токен Google."}, status=401)
        refresh = RefreshToken.for_user(User.objects.get(pk=account.id))
        return Response(
            {
                "access": str(refresh.access_token),
                "refresh": str(refresh),
                "user": {
                    "id": account.id,
                    "username": account.username,
                    "email": account.email,
                    "providers": list(account.providers),
                },
            },
            status=200,
        )


class MeView(APIView):
    permission_classes = [IsAuthenticated]

    @extend_schema(
        operation_id="auth_me",
        summary="The caller's own account",
        responses={200: AccountSerializer, 401: ErrorSerializer},
    )
    def get(self, request: Request) -> Response:
        return Response(
            {
                "id": request.user.id,
                "username": request.user.username,
                "email": request.user.email or None,
                "providers": _providers_for(request.user),
            }
        )


class LogoutView(APIView):
    """Revoke the caller's refresh token.

    The access token itself is stateless and simply expires (30 minutes), but the
    refresh token is what keeps a session alive, so blacklisting it is what
    actually ends the session. A client that no longer holds its refresh token
    still gets a success: it has nothing left to revoke.
    """

    permission_classes = [IsAuthenticated]

    @extend_schema(
        operation_id="auth_logout",
        summary="Revoke the caller's refresh token",
        request=LogoutRequestSerializer,
        # 205 even when the client no longer holds the token: there is nothing
        # left to revoke, which is not an error for the caller.
        # 400 приходит от проверки NUL на границе разбора: тело без
        # правильной формы сюда просто не доходит.
        responses={205: None, 400: ErrorSerializer, 401: ErrorSerializer},
    )
    def post(self, request: Request) -> Response:
        raw = request.data.get("refresh") if isinstance(request.data, Mapping) else None
        if raw:
            try:
                RefreshToken(raw).blacklist()
            except TokenError:
                # Already expired, already revoked, or malformed — the session is
                # gone either way, so this is not an error for the caller.
                pass
        return Response(status=status.HTTP_205_RESET_CONTENT)


class SavedSearchListView(APIView):
    permission_classes = [IsAuthenticated]

    @extend_schema(
        operation_id="saved_searches_list",
        summary="List the caller's saved searches",
        responses={200: SavedSearchSerializer(many=True), 401: ErrorSerializer},
    )
    def get(self, request: Request) -> Response:
        items = container.build_manage_saved_searches().list(request.user.id)
        return Response(SavedSearchSerializer(items, many=True).data)

    @extend_schema(
        operation_id="saved_searches_create",
        summary="Save a set of filters under a name",
        # JSON only: `filters` is a JSONField, and a form encoding delivers
        # it as a string that the field then rejects. The schema was naming
        # two media types this body cannot travel in.
        request={"application/json": SavedSearchSerializer},
        responses={
            201: SavedSearchSerializer,
            400: VALIDATION_ERROR_SCHEMA,
            401: ErrorSerializer,
        },
    )
    def post(self, request: Request) -> Response:
        serializer = SavedSearchSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        item = container.build_manage_saved_searches().create(
            owner_id=request.user.id,
            name=serializer.validated_data["name"],
            query=serializer.validated_data["query"],
            filters=serializer.validated_data.get("filters", {}),
        )
        return Response(SavedSearchSerializer(item).data, status=201)


class SavedSearchDetailView(APIView):
    permission_classes = [IsAuthenticated]

    @extend_schema(
        operation_id="saved_searches_destroy",
        summary="Delete a saved search",
        responses={204: None, 401: ErrorSerializer, 404: ErrorSerializer},
    )
    def delete(self, request: Request, saved_id: int) -> Response:
        if not container.build_manage_saved_searches().delete(request.user.id, saved_id):
            return Response({"detail": "Not found."}, status=404)
        return Response(status=204)