# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: WB Analytics API
  version: 1.0.0
  description: Контракт API сервиса аналитики товаров Wildberries. Снят с Django-реализации
    и заморожен на время миграции.
paths:
...
  /api/auth/register/:
    post:
      operationId: auth_register
      summary: Create an account
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registered'
          description: ''
        '400':
          content:
            application/json:
              schema:
                anyOf:
                - type: object
                  properties:
                    detail:
                      type: string
                  required:
                  - detail
                - type: object
                  additionalProperties:
                    oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  description: Field-keyed validation errors, as DRF renders them.
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
  /api/auth/google/:
    post:
      operationId: auth_google
      description: Exchange a Google ID token for our JWT, provisioning the account
        if new.
      summary: Exchange a Google ID token for a session
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequestRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          description: ''
        '400':
          content:
            application/json:
              schema:
                anyOf:
                - type: object
                  properties:
                    detail:
                      type: string
                  required:
                  - detail
                - type: object
                  additionalProperties:
                    oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  description: Field-keyed validation errors, as DRF renders them.
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
  /api/auth/logout/:
    post:
      operationId: auth_logout
      description: |-
        Revoke the caller's refresh token.

        The access token itself is stateless and simply expires (30 minutes), but the
        refresh token is what keeps a session alive, so blacklisting it is what
        actually ends the session. A client that no longer holds its refresh token
        still gets a success: it has nothing left to revoke.
      summary: Revoke the caller's refresh token
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LogoutRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LogoutRequestRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '205':
          description: No response body
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/auth/me/:
    get:
      operationId: auth_me
      summary: The caller's own account
      tags:
      - auth
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/saved-searches/:
    get:
      operationId: saved_searches_list
      summary: List the caller's saved searches
      tags:
      - saved-searches
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedSearch'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
    post:
      operationId: saved_searches_create
      summary: Save a set of filters under a name
      tags:
      - saved-searches
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearchRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearch'
          description: ''
        '400':
          content:
            application/json:
              schema:
                anyOf:
                - type: object
                  properties:
                    detail:
                      type: string
                  required:
                  - detail
                - type: object
                  additionalProperties:
                    oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  description: Field-keyed validation errors, as DRF renders them.
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/saved-searches/{saved_id}/:
    delete:
      operationId: saved_searches_destroy
      summary: Delete a saved search
      parameters:
      - in: path
        name: saved_id
        schema:
          type: integer
        required: true
      tags:
      - saved-searches
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
components:
  schemas:
    Account:
      type: object
      description: The caller's own account, as `/auth/me/` and Google sign-in return
        it.
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type:
          - string
          - 'null'
          format: email
        providers:
          type: array
          items:
            type: string
      required:
      - email
      - id
      - providers
      - username
...
    Error:
      type: object
      description: |-
        The `{"detail": ...}` body the shared exception handler returns.

        Not the only error body in the API: DRF answers its own serializer
        validation with a field-keyed map instead, and the handler passes that
        through untouched. The docstring here used to claim this shape covered
        "every error path", which is what a contract test disproved — see
        `FIELD_ERRORS_SCHEMA` below and `VALIDATION_ERROR_RESPONSE`.
      properties:
        detail:
          type: string
      required:
      - detail
    GoogleAuthRequestRequest:
      type: object
      properties:
        id_token:
          type: string
          minLength: 1
      required:
      - id_token
...
    LogoutRequestRequest:
      type: object
      description: |-
        `refresh` is optional: a client that has already lost it still gets 205,
        because there is nothing left to revoke.

        `allow_blank` for the same reason. Without it the schema advertises a
        minimum length the view does not enforce — an empty string is exactly the
        "already lost it" case the docstring describes, and it is answered with 205.
      properties:
        refresh:
          type: string
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 150
        password:
          type: string
          writeOnly: true
          minLength: 6
      required:
      - password
      - username
    Registered:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required:
      - id
      - username
    SavedSearch:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        query:
          type: string
          maxLength: 200
        filters:
          type: object
          additionalProperties: true
      required:
      - id
      - name
      - query
    SavedSearchRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        query:
          type: string
          minLength: 1
          maxLength: 200
        filters:
          type: object
          additionalProperties: true
      required:
      - name
      - query
...
    TokenPair:
      type: object
      properties:
        access:
          type: string
        refresh:
          type: string
        user:
          $ref: '#/components/schemas/Account'
      required:
      - access
      - refresh
      - user
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT